+++ ts/core/mixins/dom.ts
@@ -2,7 +2,7 @@ import { Modal } from 'bootstrap';
 
 export function showModal(selector: string, onShow?: () => void, onHide?: () => void) {
   const dialogEl = document.querySelector(selector);
-  new Modal(dialogEl).show();
+  Modal.getOrCreateInstance(dialogEl).show();
   onShow && dialogEl.addEventListener('shown.bs.modal', onShow);
   onHide && dialogEl.addEventListener('hide.bs.modal', onHide);
 }